Address gdkchan's comments

+++ Ryujinx.Audio/Renderer/Dsp/FixedPointHelper.cs
@@ -19,7 +19,7 @@ namespace Ryujinx.Audio.Renderer.Dsp
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static float ConvertFloat(float value, int qBits)
         {
-            return (float)value / (1 << qBits);
+            return value / (1 << qBits);
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
